requirements.txt
streamlit==1.41.0
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.0
matplotlib==3.6.2
seaborn==0.12.1
ucimlrepo==0.0.3
xgboost==1.7.6
streamlit-reveal-slides==0.1.0